+++ Sources/Nuke/Tasks/TaskFetchOriginalImage.swift
@@ -135,7 +135,7 @@ final class TaskFetchOriginalImage: AsyncPipelineTask<ImageResponse> {
 
     private func loadAsyncImage(_ fetch: @Sendable @escaping () async throws -> ImageContainer) {
         let stage = diagnostics?.beginStage(.download, queued: true)
-        operation = pipeline.configuration.dataLoadingQueue.add { [weak self] in
+        operation = pipeline.configuration.dataLoadingQueue.add(priority: priority) { [weak self] in
             await self?.performAsyncImageLoad(fetch, stage: stage)
         }
     }